Minor changes - SchemaProperty move files
